Implemented an experimental reply parser using a deque.
This is to test implementing reply buffering and parsing with something
other than a string buffer. Theoretically, avoiding the `+=` operation
performed on the string buffer is a Good Thing (ignoring for a second
the optimizations around string append in Python 2.5+).

The implementation included passes the unit tests, but needs a bit more
benchmarking, testing and polish before it should even be considered
a viable alternative to the current parsing method. It will *not* be
merged unless there are proven performance advantages.